+++ include/pfnet/constr_AC_LIN_FLOW_LIM.h
@@ -15,11 +15,26 @@
 #include "constr.h"
 #include "pfnet_config.h"
 
-#if HAVE_LINE_FLOW
-#include "line_flow.h"
-#endif
-
 #define CONSTR_AC_LIN_FLOW_LIM_INF 1e8
+
+// Line flow interface
+typedef struct LINE_FLOW_Results LINE_FLOW_Results;
+typedef struct LINE_FLOW_Params LINE_FLOW_Params;
+typedef enum { error_line_data, error_params, error_patch, infeasible, 
+	       non_binding, success, zero_limit, error_other } LINE_FLOW_Flag;
+
+LINE_FLOW_Results* LINE_FLOW_construct(double V_i_min, double V_i_max, double V_j_min, double V_j_max,
+				       double g, double b, double B_sh, double Kt_real, double Kt_shift, 
+				       double I_max_user, int flow_side,
+				       LINE_FLOW_Params* params);
+
+void LINE_FLOW_free_results(LINE_FLOW_Results* results);
+double* LINE_FLOW_get_A_matrix(LINE_FLOW_Results* results);
+double* LINE_FLOW_get_b_vector(LINE_FLOW_Results* results);
+LINE_FLOW_Flag LINE_FLOW_get_flag(LINE_FLOW_Results* results);
+int LINE_FLOW_get_number_constraints(LINE_FLOW_Results* results);
+double LINE_FLOW_get_error(LINE_FLOW_Results* results);
+char* LINE_FLOW_get_message(LINE_FLOW_Results* results);
 
 // Data
 typedef struct Constr_AC_LIN_FLOW_LIM_Data Constr_AC_LIN_FLOW_LIM_Data;