+++ python/pfnet/cbranch.pyx
@@ -18,10 +18,8 @@ class BranchError(Exception):
     """
     Branch error exception.
     """
-    def __init__(self,value):
-        self.value = value
-    def __str__(self):
-        return repr(self.value)
+
+    pass
 
 cdef class Branch:
     """