+++ python/pfnet/cfunc.pyx
@@ -17,10 +17,7 @@ class FunctionError(Exception):
     Function error exception.
     """
 
-    def __init__(self,value):
-        self.value = value
-    def __str__(self):
-        return repr(self.value)
+    pass
 
 cdef class FunctionBase:
     """