+++ python/pfnet/cload.pyx
@@ -22,10 +22,7 @@ class LoadError(Exception):
     Load error exception.
     """
 
-    def __init__(self,value):
-        self.value = value
-    def __str__(self):
-        return repr(self.value)
+    pass
 
 cdef class Load:
     """