+++ python/pfnet/cparser.pyx
@@ -15,10 +15,7 @@ class ParserError(Exception):
     Parser error exception.
     """
 
-    def __init__(self,value):
-        self.value = value
-    def __str__(self):
-        return repr(self.value)
+    pass
 
 cdef class ParserBase:
     """