+++ python/pfnet/cshunt.pyx
@@ -18,10 +18,7 @@ class ShuntError(Exception):
     Shunt error exception.
     """
 
-    def __init__(self,value):
-        self.value = value
-    def __str__(self):
-        return repr(self.value)
+    pass
 
 cdef class Shunt:
     """