+++ python/tests/test_parser.py
@@ -460,19 +460,19 @@ class TestParser(unittest.TestCase):
 
             if case.split('.')[-1] == 'mat':
                 self.assertRaises(pf.ParserError,pf.ParserART().parse,case)
-                if pf.info['raw parser']:
+                if pf.info['raw_parser']:
                     self.assertRaises(pf.ParserError,pf.ParserRAW().parse,case)
                 net = pf.ParserMAT().parse(case)
                 self.assertGreater(net.num_buses,0)
             elif case.split('.')[-1] == 'art':
                 self.assertRaises(pf.ParserError,pf.ParserMAT().parse,case)
-                if pf.info['raw parser']:
+                if pf.info['raw_parser']:
                     self.assertRaises(pf.ParserError,pf.ParserRAW().parse,case)
                 net = pf.ParserART().parse(case)
                 self.assertGreater(net.num_buses,0)
             elif case.split('.')[-1] == 'raw':
                 self.assertRaises(pf.ParserError,pf.ParserMAT().parse,case)
                 self.assertRaises(pf.ParserError,pf.ParserART().parse,case)
-                if pf.info['raw parser']:
+                if pf.info['raw_parser']:
                     net = pf.ParserRAW().parse(case)
                     self.assertGreater(net.num_buses,0)