+++ src/problem/constr/constr_AC_LIN_FLOW_LIM.c
@@ -11,12 +11,53 @@
 #include <pfnet/array.h>
 #include <pfnet/constr_AC_LIN_FLOW_LIM.h>
 
+#if !HAVE_LINE_FLOW
+
+// Dummies
+LINE_FLOW_Results* LINE_FLOW_construct(double V_i_min, double V_i_max, double V_j_min, double V_j_max,
+				       double g, double b, double B_sh, double Kt_real, double Kt_shift, 
+				       double I_max_user, int flow_side,
+				       LINE_FLOW_Params* params) {
+  return NULL;
+}
+
+void LINE_FLOW_free_results(LINE_FLOW_Results* results) {
+
+}
+
+double* LINE_FLOW_get_A_matrix(LINE_FLOW_Results* results) {
+  return NULL;
+}
+
+double* LINE_FLOW_get_b_vector(LINE_FLOW_Results* results) {
+  return NULL;
+}
+
+LINE_FLOW_Flag LINE_FLOW_get_flag(LINE_FLOW_Results* results) {
+  return error_other;
+}
+
+int LINE_FLOW_get_number_constraints(LINE_FLOW_Results* results) {
+  return 0;
+}
+
+double LINE_FLOW_get_error(LINE_FLOW_Results* results) {
+  return 0.;
+}
+
+char* LINE_FLOW_get_message(LINE_FLOW_Results* results) {
+  return "LINE_FLOW library not available";
+}
+
+#endif
+
+// Constraint data
 struct Constr_AC_LIN_FLOW_LIM_Data {
-  #if HAVE_LINE_FLOW
   LINE_FLOW_Results** results; // array of pointers to LINE_FLOW results (2 per branch and time)
   int size;
-  #endif
 };
+
+// Constraint routines
 
 Constr* CONSTR_AC_LIN_FLOW_LIM_new(Net* net) {
   Constr* c = CONSTR_new(net);
@@ -44,12 +85,10 @@ void CONSTR_AC_LIN_FLOW_LIM_init(Constr* c) {
   net = CONSTR_get_network(c);
   num = 2*NET_get_num_branches(net)*NET_get_num_periods(net); // max number of LINE_FLOW results
   data = (Constr_AC_LIN_FLOW_LIM_Data*)malloc(sizeof(Constr_AC_LIN_FLOW_LIM_Data));
-  #if HAVE_LINE_FLOW
   data->results = (LINE_FLOW_Results**)malloc(sizeof(LINE_FLOW_Results*)*num);
   data->size = num;
   for (i = 0; i < num; i++)
     data->results[i] = NULL;
-  #endif
   CONSTR_set_name(c,"linearized AC branch flow limits");
   CONSTR_set_data(c,(void*)data);
 }
@@ -63,9 +102,8 @@ void CONSTR_AC_LIN_FLOW_LIM_clear(Constr* c) {
 
 void CONSTR_AC_LIN_FLOW_LIM_count_step(Constr* c, Branch* br, int t) {
 
-  #if HAVE_LINE_FLOW
-
   // Local variables
+  int i;
   int k;
   int* G_nnz;
   int* G_row;
@@ -75,8 +113,7 @@ void CONSTR_AC_LIN_FLOW_LIM_count_step(Constr* c, Branch* br, int t) {
   REAL V_max[2];
   int offset;
   int num_constr;
-  LINE_FLOW_Results* results1;
-  LINE_FLOW_Results* results2;
+  LINE_FLOW_Results* results[2];
   Constr_AC_LIN_FLOW_LIM_Data* data;
 
   // Constr data
@@ -136,64 +173,61 @@ void CONSTR_AC_LIN_FLOW_LIM_count_step(Constr* c, Branch* br, int t) {
   }
 
   // Construct constraints Ikm <= Imax
-  results1 = LINE_FLOW_construct(V_min[0], 
-                                 V_max[0], 
-                                 V_min[1], 
-                                 V_max[1],
-                                 BRANCH_get_g(br), 
-                                 BRANCH_get_b(br),
-                                 BRANCH_get_b_k(br)+BRANCH_get_b_m(br), // total line charging
-                                 1./BRANCH_get_ratio(br,t),
-                                 BRANCH_get_phase(br,t)*180./PI, // degrees
-                                 BRANCH_get_ratingA(br),
-                                 10, 
-                                 5., 
-                                 1,     // "k" side
-                                 NULL);
+  results[0] = LINE_FLOW_construct(V_min[0], 
+				   V_max[0], 
+				   V_min[1], 
+				   V_max[1],
+				   BRANCH_get_g(br), 
+				   BRANCH_get_b(br),
+				   BRANCH_get_b_k(br)+BRANCH_get_b_m(br), // total line charging
+				   1./BRANCH_get_ratio(br,t),
+				   BRANCH_get_phase(br,t)*180./PI, // degrees
+				   BRANCH_get_ratingA(br),
+				   1,     // "k" side
+				   NULL);
 
   // Construct constraints Imk <= Imax
-  results2 = LINE_FLOW_construct(V_min[0], 
-                                 V_max[0], 
-                                 V_min[1], 
-                                 V_max[1],
-                                 BRANCH_get_g(br), 
-                                 BRANCH_get_b(br),
-				 BRANCH_get_b_k(br)+BRANCH_get_b_m(br), // total line charging
-                                 1./BRANCH_get_ratio(br,t),
-                                 BRANCH_get_phase(br,t)*180./PI, // degrees
-                                 BRANCH_get_ratingA(br),
-                                 10, 
-                                 5.,    
-                                 2,     // "m" side
-                                 NULL);
+  results[1] = LINE_FLOW_construct(V_min[0], 
+				   V_max[0], 
+				   V_min[1], 
+				   V_max[1],
+				   BRANCH_get_g(br), 
+				   BRANCH_get_b(br),
+				   BRANCH_get_b_k(br)+BRANCH_get_b_m(br), // total line charging
+				   1./BRANCH_get_ratio(br,t),
+				   BRANCH_get_phase(br,t)*180./PI, // degrees
+				   BRANCH_get_ratingA(br),
+				   2,     // "m" side
+				   NULL);
 
   // Save results
-  data->results[2*BRANCH_get_index(br)+0+offset] = results1; 
-  data->results[2*BRANCH_get_index(br)+1+offset] = results2;
- 
-  // Check results
-  if (LINE_FLOW_get_flag(results1) != 2) {
-    CONSTR_set_error(c,LINE_FLOW_get_message(results1));    
-    return;
-  }
-  if (LINE_FLOW_get_flag(results2) != 2) {
-    CONSTR_set_error(c,LINE_FLOW_get_message(results2));
-    return;
+  for (i = 0; i < 2; i++) {
+    if (data->results[2*BRANCH_get_index(br)+i+offset])
+      LINE_FLOW_free_results(data->results[2*BRANCH_get_index(br)+i+offset]); // delete existing
+    data->results[2*BRANCH_get_index(br)+i+offset] = results[i];              // save new
   }
  
-  // Count
-  num_constr = LINE_FLOW_get_number_constraints(results1)+LINE_FLOW_get_number_constraints(results2);
-  (*G_row) += num_constr;
-  for (k = 0; k < 2; k++) {
-    if (BUS_has_flags(bus[k],FLAG_VARS,BUS_VAR_VMAG)) // v_mag variable
-      (*G_nnz) += num_constr;
-    if (BUS_has_flags(bus[k],FLAG_VARS,BUS_VAR_VANG)) // v_ang variable
-      (*G_nnz) += num_constr;
+  // Check success
+  for (i = 0; i < 2; i++) {
+    if (LINE_FLOW_get_flag(results[i]) != success && LINE_FLOW_get_flag(results[i]) != non_binding) {
+      CONSTR_set_error(c,LINE_FLOW_get_message(results[i]));
+      return;
+    }
   }
-
-  #else
-  CONSTR_set_error(c,"unable to construct AC_LIN_FLOW_LIM constraint - line_flow library not available");
-  #endif
+ 
+  // Count rows and nnz
+  for (i = 0; i < 2; i++) {
+    if (LINE_FLOW_get_flag(results[i]) == success) {
+      num_constr = LINE_FLOW_get_number_constraints(results[i]);
+      for (k = 0; k < 2; k++) {
+	if (BUS_has_flags(bus[k],FLAG_VARS,BUS_VAR_VMAG)) // v_mag variable
+	  (*G_nnz) += num_constr;
+	if (BUS_has_flags(bus[k],FLAG_VARS,BUS_VAR_VANG)) // v_ang variable
+	  (*G_nnz) += num_constr;
+      }
+      (*G_row) += num_constr;
+    }
+  }
 }
 
 void CONSTR_AC_LIN_FLOW_LIM_allocate(Constr* c) {
@@ -209,15 +243,15 @@ void CONSTR_AC_LIN_FLOW_LIM_allocate(Constr* c) {
   G_nnz = CONSTR_get_G_nnz(c);
   G_row = CONSTR_get_G_row(c);
 
-  // J f
+  // J f (from f(x) = 0)
   CONSTR_set_J(c,MAT_new(0,num_vars,0));
   CONSTR_set_f(c,VEC_new(0));
 
-  // A b
+  // A b (from Ax = b)
   CONSTR_set_A(c,MAT_new(0,num_vars,0));
   CONSTR_set_b(c,VEC_new(0));
 
-  // G l u
+  // G l u (from l <= Gx <= u)
   CONSTR_set_G(c,MAT_new(G_row,    // size1 (rows)
 			 num_vars, // size2 (cols)
 			 G_nnz));  // nnz
@@ -226,8 +260,6 @@ void CONSTR_AC_LIN_FLOW_LIM_allocate(Constr* c) {
 }
 
 void CONSTR_AC_LIN_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t) {
-
-  #if HAVE_LINE_FLOW
 
   // Local variables
   Mat* G;
@@ -249,6 +281,9 @@ void CONSTR_AC_LIN_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t) {
 
   // Constr data
   net = CONSTR_get_network(c);
+  G = CONSTR_get_G(c);
+  l = CONSTR_get_l(c);
+  u = CONSTR_get_u(c);
   G_nnz = CONSTR_get_G_nnz_ptr(c);
   G_row = CONSTR_get_G_row_ptr(c);
   data = (Constr_AC_LIN_FLOW_LIM_Data*)CONSTR_get_data(c);
@@ -305,46 +340,51 @@ void CONSTR_AC_LIN_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t) {
   // Construct l <= Gx <= u
   for (i = 0; i < 2; i++) {
 
-    num_constr = LINE_FLOW_get_number_constraints(results[i]);
-    A = LINE_FLOW_get_A_matrix(results[i]);
-    b = LINE_FLOW_get_b_vector(results[i]);
+    if (LINE_FLOW_get_flag(results[i]) == success) {
+
+      num_constr = LINE_FLOW_get_number_constraints(results[i]);
+      A = LINE_FLOW_get_A_matrix(results[i]);
+      b = LINE_FLOW_get_b_vector(results[i]);
     
-    for (j = 0; j < num_constr; j++) {
+      for (j = 0; j < num_constr; j++) {
   
-      // l and u
-      VEC_set(l,*G_row,-CONSTR_AC_LIN_FLOW_LIM_INF); // lower limit
-      VEC_set(u,*G_row,b[j]);                        // upper limit
-      
-      for (k = 0; k < 2; k++) {
+	// l and u
+	VEC_set(l,*G_row,-CONSTR_AC_LIN_FLOW_LIM_INF); // lower limit (minus infinity)
+	VEC_set(u,*G_row,b[j]);                        // upper limit
+	
+	for (k = 0; k < 2; k++) {
   
-        if (BUS_has_flags(bus[k],FLAG_VARS,BUS_VAR_VMAG)) { // v_mag variable
+	  if (BUS_has_flags(bus[k],FLAG_VARS,BUS_VAR_VMAG)) { // v_mag variable
 
-	  // G
-          MAT_set_i(G,*G_nnz,*G_row);
-          MAT_set_j(G,*G_nnz,BUS_get_index_v_mag(bus[k],t)); // vk time t
-          MAT_set_d(G,*G_nnz,A[3*j+k]);
-          (*G_nnz)++;
-        }
-
-	if (BUS_has_flags(bus[k],FLAG_VARS,BUS_VAR_VANG)) { // v_ang variable
-
-          // G
-          MAT_set_i(G,*G_nnz,*G_row);
-          MAT_set_j(G,*G_nnz,BUS_get_index_v_ang(bus[k],t)); // wk time t
-	  if (k == 0) 
-	    MAT_set_d(G,*G_nnz,A[3*j+2]);
-	  else
-	    MAT_set_d(G,*G_nnz,-A[3*j+2]);
-          (*G_nnz)++;
-        }
+	    // G
+	    MAT_set_i(G,*G_nnz,*G_row);
+	    MAT_set_j(G,*G_nnz,BUS_get_index_v_mag(bus[k],t)); // vk time t
+	    MAT_set_d(G,*G_nnz,A[3*j+k]);                      // k = 0 gives Vi, k = 1 gives Vj
+	    (*G_nnz)++;
+	  }
+	  
+	  if (BUS_has_flags(bus[k],FLAG_VARS,BUS_VAR_VANG)) { // v_ang variable
+	    
+	    // G
+	    MAT_set_i(G,*G_nnz,*G_row);
+	    MAT_set_j(G,*G_nnz,BUS_get_index_v_ang(bus[k],t)); // wk time t
+	    if (k == 0) 
+	      MAT_set_d(G,*G_nnz,A[3*j+2]);                    // theta_i
+	    else
+	      MAT_set_d(G,*G_nnz,-A[3*j+2]);                   // theta_j
+	    (*G_nnz)++;
+	  }
+	  else {
+	    if (k == 0)
+	      VEC_add_to_entry(u,*G_row,-A[3*j+2]*BUS_get_v_ang(bus[k],t));
+	    else
+	      VEC_add_to_entry(u,*G_row,A[3*j+2]*BUS_get_v_ang(bus[k],t));
+	  }
+	}
+	(*G_row)++;
       }
-      (*G_row)++;
     }
   }
-  
-  #else
-  CONSTR_set_error(c,"unable to construct AC_LIN_FLOW_LIM constraint - line_flow library not available");
-  #endif  
 }
 
 void CONSTR_AC_LIN_FLOW_LIM_eval_step(Constr* c, Branch* br, int t, Vec* values, Vec* values_extra) {
@@ -366,13 +406,11 @@ void CONSTR_AC_LIN_FLOW_LIM_free(Constr* c) {
 
   // Free
   if (data) {
-    #if HAVE_LINE_FLOW
     for (i = 0; i < data->size; i++) {
       if (data->results[i])
 	LINE_FLOW_free_results(data->results[i]);
     }
     free(data->results);
-    #endif
     free(data);
   }