+++ src/problem/problem.c
@@ -113,9 +113,24 @@ void PROB_analyze(Prob* p) {
   // Count
   for (t = 0; t < NET_get_num_periods(p->net); t++) {
     for (k = 0; k < NET_get_num_branches(p->net); k++) {
+      
       br = NET_get_branch(p->net,k);
+      
+      // Constraints
       CONSTR_list_count_step(p->constr,br,t);
+      if (CONSTR_list_has_error(p->constr)) {
+	strcpy(p->error_string,CONSTR_list_get_error_string(p->constr));
+	p->error_flag = TRUE;
+	return;
+      }
+      
+      // Functions
       FUNC_list_count_step(p->func,br,t);
+      if (FUNC_list_has_error(p->func)) {
+	strcpy(p->error_string,FUNC_list_get_error_string(p->func));
+	p->error_flag = TRUE;
+	return;
+      }
     }
   }
 
@@ -136,9 +151,24 @@ void PROB_analyze(Prob* p) {
   // Analyze
   for (t = 0; t < NET_get_num_periods(p->net); t++) {
     for (k = 0; k < NET_get_num_branches(p->net); k++) {
+      
       br = NET_get_branch(p->net,k);
+      
+      // Constraints
       CONSTR_list_analyze_step(p->constr,br,t);
+      if (CONSTR_list_has_error(p->constr)) {
+	strcpy(p->error_string,CONSTR_list_get_error_string(p->constr));
+	p->error_flag = TRUE;
+	return;
+      }
+
+      // Functions
       FUNC_list_analyze_step(p->func,br,t);
+      if (FUNC_list_has_error(p->func)) {
+	strcpy(p->error_string,FUNC_list_get_error_string(p->func));
+	p->error_flag = TRUE;
+	return;
+      }
     }
   }
 
@@ -218,8 +248,19 @@ void PROB_apply_heuristics(Prob* p, Vec* point) {
 
 void PROB_clear_error(Prob* p) {
   if (p) {
+
+    // Problem
     p->error_flag = FALSE;
     strcpy(p->error_string,"");
+
+    // Constraints
+    CONSTR_list_clear_error(p->constr);
+
+    // Functions
+    FUNC_list_clear_error(p->func);
+    
+    // Network
+    NET_clear_error(p->net);
   }
 }
 
@@ -259,10 +300,32 @@ void PROB_eval(Prob* p, Vec* point) {
   // Eval
   for (t = 0; t < NET_get_num_periods(p->net); t++) {
     for (k = 0; k < NET_get_num_branches(p->net); k++) {
+    
       br = NET_get_branch(p->net,k);
+      
+      // Constraints
       CONSTR_list_eval_step(p->constr,br,t,x,y);
+      if (CONSTR_list_has_error(p->constr)) {
+	strcpy(p->error_string,CONSTR_list_get_error_string(p->constr));
+	p->error_flag = TRUE;
+	return;
+      }
+      
+      // Functions
       FUNC_list_eval_step(p->func,br,t,x);
+      if (FUNC_list_has_error(p->func)) {
+	strcpy(p->error_string,FUNC_list_get_error_string(p->func));
+	p->error_flag = TRUE;
+	return;
+      }
+      
+      // Network
       NET_update_properties_step(p->net,br,t,x);
+      if (NET_has_error(p->net)) {
+	strcpy(p->error_string,NET_get_error_string(p->net));
+	p->error_flag = TRUE;
+	return;
+      }
     }
   }
 
@@ -297,8 +360,16 @@ void PROB_store_sens(Prob* p, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl) {
   // Store sens
   for (t = 0; t < NET_get_num_periods(p->net); t++) {
     for (i = 0; i < NET_get_num_branches(p->net); i++) {
+
       br = NET_get_branch(p->net,i);
+      
+      // Constraints
       CONSTR_list_store_sens_step(p->constr,br,t,sA,sf,sGu,sGl);
+      if (CONSTR_list_has_error(p->constr)) {
+	strcpy(p->error_string,CONSTR_list_get_error_string(p->constr));
+	p->error_flag = TRUE;
+	return;
+      }
     }
   }
 }